adding routes for create script
