+++ server/routes/skilltest/script.routes.js
@@ -1,8 +1,10 @@
 const router = require('express').Router(),
-  skillTestController = require('./../../controllers/skilltest.controller.js');
+  skillTestController = require('./../../controllers/skilltest.controller'),
+  converterService = require('./../../modules/skilltest/converter.service'),
+  config = require('./../../config/skilltest.config');
 
 /**
- * Get list of templates
+ * Get list of scripts
  * Query params: app
  * Return Type: Array
  */
@@ -16,36 +18,66 @@ router.get('/', (req, res) => {
     }, (error) => {
       res.send(error);
     });
-
-  /**
-   * Sample:
-   * [{
-      "uuid": "sample-uuid-123",
-      "name": "Move Cell Content",
-      "meta": {
-        "version": 1,
-        "description": "Test template for skill related to move cell content",
-        "skill": "move cell content",
-        "app": "excel"
-        }
-    }]
-   */
-
 });
 
 /**
- * Get template by id
- * Query params: app
- * Return Type: Array
+ * Get script by id
+ * Query params: app, format
+ * Return Type: Object
  */
 
 router.get('/:sleId', (req, res) => {
   let appType   = req.query.app,
-    sleId  = req.params.sleId;
+      format = req.query.format,
+      sleId  = req.params.sleId;
+
 
   skillTestController.getScriptBySleId(sleId, appType)
     .then((scripts) => {
-      res.send(scripts);
+
+      if(!scripts.length) {
+        res.status(404).send(config.messages.notFound);
+      } else {
+
+        if (format === 'xml') {
+          let xmlContent = converterService.jsonToDistXml(scripts[0]);
+          res.set('Content-Type', 'text/xml');
+          res.send(xmlContent);
+        } else {
+          res.send(scripts[0]);
+        }
+      }
+
+    }, (error) => {
+      res.send(error);
+    });
+});
+
+/**
+ * Get script pathways by id
+ * Query params: app
+ * Return Type: Text
+ */
+
+router.get('/:sleId/pathways', (req, res) => {
+  let appType   = req.query.app,
+    sleId  = req.params.sleId;
+
+
+  skillTestController.getScriptBySleId(sleId, appType)
+    .then((scripts) => {
+
+      if(!scripts.length) {
+        res.status(404).send(config.messages.notFound);
+      } else {
+        let javaContent = converterService.jsonToDistJava(scripts[0]);
+
+        res.set('Content-type', 'text/plain');
+        res.charset = 'UTF-8';
+
+        res.send(javaContent);
+      }
+
     }, (error) => {
       res.send(error);
     });
